requirements.txt
fastapi==0.116.0
langchain>=0.3.7
langchain-community>=0.3.7
langchain-core>=0.3.15
langchain-ollama>=0.3.2
chromadb==0.4.24
python-dotenv==1.0.1
requests==2.31.0
streamlit==1.28.0
pydantic<3.0.0,>=1.8.0
